test trimmed code
